fix: 解决网站域名可以重复添加的问题
