+++ backend/i18n/lang/en.yaml
@@ -75,6 +75,7 @@ ErrBackupMatch: 'the backup file does not match the current partial data of the 
 ErrBackupExist: 'the backup file corresponds to a portion of the original data that does not exist: {{ .detail}}"'
 ErrPHPResource: 'The local runtime does not support switching！'
 ErrPathPermission: 'A folder with non-1000:1000 permissions was detected in the index directory, which may cause Access denied errors when accessing the website.'
+ErrDomainIsUsed: "Domain is already used by website {{ .name }}"
 
 #ssl
 ErrSSLCannotDelete: "The certificate is being used by the website and cannot be removed"
