+++ backend/i18n/lang/zh.yaml
@@ -75,6 +75,7 @@ ErrBackupMatch: '该备份文件与当前网站部分数据不匹配 {{ .detail}
 ErrBackupExist: '该备份文件对应部分源数据不存在 {{ .detail}}"'
 ErrPHPResource: '本地运行环境不支持切换！'
 ErrPathPermission: 'index 目录下检测到非 1000:1000 权限文件夹，可能导致网站访问 Access denied 错误'
+ErrDomainIsUsed: "域名已被网站【{{ .name }}】使用"
 
 #ssl
 ErrSSLCannotDelete: "证书正在被网站使用，无法删除"
